Added an "asteroid belt" of cubes. Wireframe state can be toggled.
"Asteroid belt" of cubes isn't exactly perfect because of how the rand() function behaves - looking for a way so 100 cubes can be visible. Left arrow key can be held to make the asteroid belt move in the opposite direction.  Up arrow key can be pressed to change cubes to wireframe state. Down arrow key can be pressed to change cubes to solid state.